Splitting front end js out to an included file.
This may or may not work out - so long as babel doesn't need to compile anything
in the code I think it will work. Might be too fragile but we'll try it for now.
