+++ templates/index.js
@@ -1,4 +1,5 @@
 import svgDimensions from '../lib/svgDimensions';
+import bootstrapApp from '../public/index';
 
 const stylesheetTag = href => (href ? `<link rel="stylesheet" href="${href}" />` : '');
 
@@ -114,104 +115,11 @@ export default ({stylesheetLink, customCSS, currentUser, matchingUsers}) => {
     <script src="https://unpkg.com/jwt-decode@2.2.0/build/jwt-decode.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/qs/6.5.0/qs.min.js"></script>
     <script type="text/javascript">
-      var params = window.Qs.parse(window.location.search, { ignoreQueryPrefix: true });
       var currentUser = ${JSON.stringify(currentUser)};
       var matchingUsers = ${JSON.stringify(matchingUsers)};
+      var bootstrapApp = ${bootstrapApp.toString()};
 
-      try {
-        loadLinkPage(window.jwt_decode(params.child_token));
-      } catch(e) {
-        console.error(e);
-        loadInvalidTokenPage();
-      }
-
-      function loadLinkPage(token) {
-        var linkEl = document.getElementById('link');
-        var skipEl = document.getElementById('skip');
-        var connections = matchingUsers.reduce(function(acc, user) {
-          return acc.concat(user.connections);
-        }, []);
-
-        var authorize = function(domain, qs) {
-          var query = Object.keys(qs)
-            .filter(function(key) {
-              return !!qs[key];
-            })
-            .map(function (key) {
-              return key + '=' + encodeURI(qs[key]);
-            }).join('&');
-
-          window.location = domain + 'authorize?' + query;
-        };
-
-        var updateContinueUrl = function(linkEl, domain, state) {
-          linkEl.href = domain + 'continue?state=' + state;
-        };
-
-        linkEl.addEventListener('click', function(e) {
-          authorize(token.iss, {
-            client_id: params.client_id,
-            redirect_uri: params.redirect_uri,
-            response_type: params.response_type,
-            scope: params.scope,
-            state: params.original_state,
-            nonce: params.nonce,
-            link_account_token: params.child_token,
-            prevent_sign_up: true,
-            connection: connections[0]
-          });
-        });
-
-        updateContinueUrl(skipEl, token.iss, params.state);
-
-        if (params.error_type === 'accountMismatch') {
-          loydAccountMismatchError();
-        }
-      }
-
-      function loadInvalidTokenPage() {
-        var containerEl = document.getElementById('content-container');
-        var labelEl = document.getElementById('label-value');
-        var linkEl = document.getElementById('link');
-
-        containerEl.innerHTML = '';
-        containerEl.appendChild(
-          el('div', {}, [
-            el('p', {}, [
-              text('You seem to have reached this page in error. Please try logging in again')
-            ])
-          ])
-        );
-
-        linkEl.disabled = true;
-      };
-
-      function loadAccountMismatchError() {
-        var messageEl = document.getElementById('error-message');
-        var msg = "Accounts must have matching email addresses. Please try again.";
-
-        messageEl.innerHTML = msg;
-        messageEl.style.display = 'block';
-      };
-
-      function el(tagName, attrs, childEls) {
-        var element = document.createElement(tagName);
-        var children = childEls || [];
-
-        for (var i in Object.keys(attrs || {})) {
-          element.setAttribute(i, attrs[i]);
-        }
-
-        for (var i in children) {
-          element.appendChild(children[i]);
-        }
-
-        return element;
-      }
-
-      function text(content) {
-        return document.createTextNode(content);
-      }
+      bootstrapApp(currentUser, matchingUsers);
     </script>
   </body>
 </html>
